pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pygarble"
version = "0.1.6"
description = "A Python package for detecting garbled text using multiple detection strategies with a scikit-learn-like interface"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ujjwal", email = "ujjwal@example.com"},
]
keywords = ["garble", "text", "detection", "nlp", "machine-learning", "scikit-learn"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fasttext-wheel>=0.9.2",
    "numpy>=1.21.0",
    "pyspellchecker>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]

[project.urls]
Homepage = "https://github.com/brightertiger/pygarble"
Repository = "https://github.com/brightertiger/pygarble"
Issues = "https://github.com/brightertiger/pygarble/issues"
Documentation = "https://github.com/brightertiger/pygarble#readme"

[tool.black]
line-length = 79
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.flake8]
max-line-length = 79
extend-ignore = ["E203", "W503"]

[tool.isort]
profile = "black"
line_length = 79
